+++ util.cpp
@@ -4,15 +4,15 @@ struct Context context;
 
 Vectorf util::intersection(std::pair<Vectorf, Vectorf> a, std::pair<Vectorf, Vectorf> b)
 {
-	double a1 = a.second.y - a.first.y;
-	double b1 = a.first.x - a.second.x;
-	double c1 = a1 * (a.first.x) + b1 * (a.first.y);
+	float a1 = a.second.y - a.first.y;
+	float b1 = a.first.x - a.second.x;
+	float c1 = a1 * (a.first.x) + b1 * (a.first.y);
 
-	double a2 = b.second.y - b.first.y;
-	double b2 = b.first.x - b.second.x;
-	double c2 = a2 * (b.first.x) + b2 * (b.first.y);
+	float a2 = b.second.y - b.first.y;
+	float b2 = b.first.x - b.second.x;
+	float c2 = a2 * (b.first.x) + b2 * (b.first.y);
 
-	double determinant = a1 * b2 - a2 * b1;
+	float determinant = a1 * b2 - a2 * b1;
 
 	if (determinant == 0)
 	{
@@ -20,8 +20,8 @@ Vectorf util::intersection(std::pair<Vectorf, Vectorf> a, std::pair<Vectorf, Vec
 	}
 	else
 	{
-		double x = (b2 * c1 - b1 * c2) / determinant;
-		double y = (a1 * c2 - a2 * c1) / determinant;
+		float x = (b2 * c1 - b1 * c2) / determinant;
+		float y = (a1 * c2 - a2 * c1) / determinant;
 		if (x >= std::min(a.first.x, a.second.x) && x <= std::max(a.first.x, a.second.x) &&
 			x >= std::min(b.first.x, b.second.x) && x <= std::max(b.first.x, b.second.x) &&
 			y >= std::min(a.first.y, a.second.y) && y <= std::max(a.first.y, a.second.y) &&
